// API_Client/client/models/company.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Online Shop API
 * REST API for the online shop backend
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface Company {
    'id': string;
    'name': string;
    'description'?: string;
    'logoUrl'?: string;
    'isActive': boolean;
    'sortOrder': number;
    'createdAt': string;
    'updatedAt': string;
}